+++ NEWS.md
@@ -13,6 +13,7 @@
 * Updated intro vignette with new functionality
 * Added a new case studies vignette section
 * Added a case study exploring global trends in Tuberculosis incidence rates
+* Added a work around using `utils::read.csv` for downloading the data when `fread::data.table` fails. Related to this [issue](https://github.com/seabbs/getTBinR/issues/18).
 
 # getTBinR 0.5.0.0